ai_screenwriter_tool/brainstorming_assistance.py
```python
import random

class BrainstormingAssistance:
    def __init__(self, user, project):
        self.user = user
        self.project = project

    def provideBrainstormingAssistance(self):
        # Generate random prompts for brainstorming
        prompts = ["What if the main character had a secret?", 
                   "What if there was a plot twist in the middle of the story?", 
                   "What if the villain was not who it seems?", 
                   "What if the story was set in a different time or place?"]

        prompt = random.choice(prompts)
        return prompt

    def createStoryStructure(self):
        # Generate a basic story structure
        structure = ["Introduction", "Rising Action", "Climax", "Falling Action", "Resolution"]
        return structure
```