ai_screenwriter_tool/tests/test_brainstorming_assistance.py
```python
import unittest
from ai_screenwriter_tool.brainstorming_assistance import provideBrainstormingAssistance

class TestBrainstormingAssistance(unittest.TestCase):

    def setUp(self):
        self.user = {
            "name": "Test User",
            "email": "testuser@example.com",
            "subscription": "free"
        }
        self.project = {
            "title": "Test Project",
            "genre": "Drama",
            "outline": "A test project for unit testing."
        }

    def test_provideBrainstormingAssistance(self):
        result = provideBrainstormingAssistance(self.user, self.project)
        self.assertIsNotNone(result)
        self.assertIsInstance(result, dict)
        self.assertIn('ideas', result)
        self.assertIn('structure', result)

if __name__ == '__main__':
    unittest.main()
```